requirements.txt
streamlit==1.52.2
tensorflow==2.15.0
numpy==2.4.0
pillow==12.0.0